// src/types/popupTypes.ts
export type PopupDirection = 'n' | 'e' | 'w' | 's';
export type PopupPositionType = 'min' | 'center' | 'max';
// ...
export interface PopupProps {
  arrowDistance?: number;
  arrowHeight?: number;
  arrowOffset?: number;
  arrowPosition?: PopupPositionType;
  arrowWidth?: number;
  class?: string;
  cornerRadius?: number;
  debugMode?: boolean;
  direction?: PopupDirection;
  eventOff?: keyof HTMLElementEventMap;
  eventOn?: keyof HTMLElementEventMap;
  isManual?: boolean;
  isModal?: boolean;
  isPositioned?: boolean;
  isTooltip?: boolean;
  logging?: boolean;
  paddingX?: number;
  paddingY?: number;
  popupOffset?: number;
  popupPosition?: PopupPositionType;
  positioner?: string;
  shiftX?: number;
  shiftY?: number;
  zIndex?: number;
}
// ...
export interface PopupDirectionMapItem {
  yReverse: boolean;
  switchXY: boolean;
  svgElements: string[];
  startOffset: 'H' | 'W' | '-H' | '-W' | '0';
}

export type PopupDirectionMap = {
  [key in PopupDirection]: PopupDirectionMapItem;
};

export interface PopupGlobalSettings {
  log: boolean;
  debugMode: boolean;
  modes: {
    [key in PopupModeKeys]: {
      props: PopupProps;
      directionMap?: PopupDirectionMap;
    };
  };
  svgElements: {
    [key: string]: {
      fill?: string;
      pathTemplate: string;
    };
  };
}